backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
psutil>=5.9.0
pydantic>=2.0.0
httpx>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
asyncssh>=2.14.0
APScheduler>=3.10.4
cryptography>=41.0.0 # For SSH Key Encryption at Rest
pywebpush>=1.2.0 # For Web Push Notifications
python-jose[cryptography]>=3.3.0 # For JWT tokens
passlib>=1.7.4 # For password hashing
bcrypt==4.1.3 # Pin to 4.x for passlib compatibility
python-multipart>=0.0.6 # For OAuth2 form data handling
importlib-metadata>=6.0.0 # For plugin entry points (Python <3.10 compat)
pymysql>=1.1.0  # MySQL metrics collection
pytest>=7.4.0  # Testing framework
pytest-asyncio>=0.21.0  # Async test support
alembic>=1.12.0  # Database migrations
pydantic-settings>=2.0.0  # Configuration management

# Additional dependencies for Run 3
docker>=7.0.0  # Docker plugin support
pymongo>=4.0.0  # MongoDB plugin support
influxdb-client>=1.40.0  # InfluxDB plugin support
pyyaml>=6.0
redis>=5.0.0  # Session storage and caching
apprise>=1.7.0  # Notification gateway (78+ services)
authlib>=1.3.0  # OAuth2 client library
